requirements_dev.txt
-r requirements.txt
httpx
pytest
python-dotenv
mkdocs-material
mkdocstrings
